refactor: use the traced file copy from OpenNext
