+++ packages/cloudflare/package.json
@@ -73,7 +73,7 @@
   "dependencies": {
     "@ast-grep/napi": "^0.34.1",
     "@dotenvx/dotenvx": "catalog:",
-    "@opennextjs/aws": "https://pkg.pr.new/@opennextjs/aws@704",
+    "@opennextjs/aws": "https://pkg.pr.new/@opennextjs/aws@712",
     "enquirer": "^2.4.1",
     "glob": "catalog:",
     "ts-morph": "catalog:",
